fix{step2}: fix wrong observation and using GLMM
correct the section trait x accuracy as the original one using trial level (it supposed to be individual level).

use GLMM to fit how condition and context affect correct selection now. (as it has higher power) It lead to some new significant difference.

compare with and without outlier in terms of using traits to predict lgRT.